requirements.txt
asgiref==3.3.1
boto3==1.16.43
botocore==1.19.43
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3.1
Django==3.1.4
django-cors-headers==3.6.0
django-environ==0.4.5
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
PyJWT==2.0.0
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
